+++ src/common/utils/xml_parser.h
@@ -1,4 +1,5 @@
-#pragma once
+#ifndef CORE_COMMON_UTILS_XML_PARSER_H
+#define CORE_COMMON_UTILS_XML_PARSER_H
 
 #include <boost/property_tree/ptree.hpp>
 #include <boost/property_tree/xml_parser.hpp>
@@ -42,4 +43,7 @@ private:
         }
     };
 };
+
 } // namespace uh::cluster
+
+#endif